backend/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
pymongo>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
google-generativeai>=0.8.0
python-engineio>=4.9.0
python-socketio>=5.11.0
Werkzeug>=3.0.0
hypothesis>=6.92.0
pytest>=7.4.0
